added a recipe for the interior door, and added a new setFuel command because it was getting super annoying to not have infinite fuel :/

+++ src/main/java/mdteam/ait/tardis/Tardis.java
@@ -115,7 +115,8 @@ public class Tardis {
     public double getFuel() {
         return this.getHandlers().getFuel().getFuel();
     }
-    public void setFuelCount(int i) {
+
+    public void setFuelCount(double i) {
         this.getHandlers().getFuel().setFuelCount(i);
     }
     public boolean isRefueling() {
